Updated(): JWT revocation strategy
