+++ backend/db/migrate/20220303175933_devise_create_users.rb
@@ -8,8 +8,6 @@ class DeviseCreateUsers < ActiveRecord::Migration[6.1]
 
       t.string :wallet_address, null: false, default: ''
       t.string :encrypted_password, null: false, default: ''
-      t.string :jti, null: false
-      t.string :signature, null: false
 
       t.timestamps null: false
     end
