+++ backend/db/seeds/00_user.rb
@@ -4,10 +4,8 @@
 # The data can then be loaded with the bin/rails db:seed command (or created alongside the database with db:setup).
 
 test_user = User.find_or_initialize_by(wallet_address: '0xbD6e2111fa9ea5B70D9F2832925391031Ce275f4')
+binding.pry;
 test_user.update!(
-  {
-    wallet_address: '0xbD6e2111fa9ea5B70D9F2832925391031Ce275f4',
-    signature: '0xnjqnejkqwnekwqenqnelqwenwqlenqwlkenqwlenqwleqw'
-  }
+  { wallet_address: '0xbD6e2111fa9ea5B70D9F2832925391031Ce275f4' }
 )
 p 'A User sample user have been seeded'
